requirements.txt
# Core libraries
langchain>=0.1.16
langchain-core>=0.1.45
langchain-community>=0.0.34
langchain-pinecone>=0.0.6
openai>=1.33.0

# Pinecone client
pinecone-client>=3.2.2

# Streamlit for frontend
streamlit>=1.35.0

# Web scraping and parsing
beautifulsoup4>=4.12.3
requests>=2.31.0

# PDF support (optional for uploaded docs)
PyMuPDF>=1.23.25  # fitz
pdfminer.six>=20221105

# Environment variable loading
python-dotenv>=1.0.1

# Optional for performance/logging/debugging
tqdm>=4.66.4
httpx>=0.27.0
loguru>=0.7.2
